Add support for remote downloading
Summary:
	*gdb/kvx-target.c (struct kvx_target): Add the m_local_connection
	attribute to the struct
	(kvx_target_attach): Set m_local_connection and remove the no
	executable warning hack
	(kvx_target::mourn_inferior): Check the stratum stack before unpush
	(kvx_change_file): Adapt the code for the remote executable
	*gdb/kvx-tdep.c (sync_insert_remove_breakpoint): Fix cast warning
	*gdb/solib-kvx-bare.c (kvx_bare_check_so): Adapt the code for remote
	shared libraries.
	(kvx_bare_solib_current_sos): Add the solib_search_path to the so
	path if solib_search_path begins with 'target:/' and the so path
	is not absolute
	*gdb/solib.c: Make solib_search_path non static
	*gdb/solib.h: Likewise

Ref T21107

Test Plan: CI

Reviewers: O12 gdb!

Maniphest Tasks: T21107

Differential Revision: https://phab.kalray.eu/D28315

+++ gdb/kvx-tdep.c
@@ -997,7 +997,7 @@ sync_insert_remove_breakpoint (CORE_ADDR addr, int len, uint32_t value)
 
   buf = (char *) malloc (size);
   sprintf (buf, "Qkalray.unified_write:%" PRIx64 ",%d:%" PRIx64, addr, len,
-	   (unsigned long long) value);
+	   (unsigned long) value);
   putpkt (rt, buf);
   getpkt (rt, &buf, &size, 0);
   free (buf);
